requirements.txt
discord.py
fuzzywuzzy[speedup]
Flask